+++ src/components/user/AddUser/AddUser.css
@@ -1,0 +1,30 @@
+input{
+    display: flex;
+    width: 350px;
+    height: 30px;
+    margin-top: 10px;
+    margin-left: 10px;
+    border-radius: 15px;
+    border-width: 0px;
+}
+::placeholder{
+    font-size: 15px;
+    padding-right: 10px;
+}
+.id_search{
+    padding-right: 30px;
+    margin-top: 50px;
+}
+
+.id_search button{
+    color:white;
+    background-color: rgb(0, 165, 146);
+    font-size: 15px;
+    font-weight: bold;
+    width: 90px;
+    height: 30px;
+    margin-top: 10px;
+    border-radius: 15px;
+    border-width: 0px;
+    margin-right: 10px;
+}
